Add files via upload
A Ride class is defined to simulate the operation of a ride. The class contains several properties such as facility name, open state, maximum capacity, operator, wait queue and history. It implements the methods of the interface RideInterface to provide functions such as queuing, removing, running cycles, adding and querying history for visitors. In addition, the class supports importing and exporting history to files, handling parsing of visitor information, and error handling.