Add files via upload
VisitorComparator class implements the Comparator<Visitor> interface, which is used to sort visitors according to their ticket type and visit duration. Firstly, the comparison is done according to the ticket type (VIP is preferred to Normal); if the ticket types are the same, then further ascending order is done according to the length of visit, in order to achieve the customised sorting rules with multiple criteria.